+++ R/excel_helpers.R
@@ -1989,8 +1989,6 @@ modify_number_formats <- function(formats_to_modify, ...){
 #' @param subheader_underline FALSE by default. If TRUE underlines the subheaders.
 #' @param colored_tabs FALSE by default. If TRUE colors the tabs according to the
 #' subheader colors.
-#' @param file If NULL, opens the output as temporary file. If a filename with path
-#' is specified, saves the output to the specified path.
 #' @param print TRUE by default. If TRUE prints the output, if FALSE doesn't print
 #' anything. Can be used if one only wants to catch the combined workbook.
 #'
@@ -2064,13 +2062,26 @@ modify_number_formats <- function(formats_to_modify, ...){
 #'                          style               = excel_output_style(toc_header_font_size = 20,
 #'                                                                   toc_other_font_size = 14))
 #'
+#' # To save a table as xlsx file you have to set the path and filename in the
+#' # style element
+#' # Example files paths
+#' workbook_file <- tempfile(fileext = ".xlsx")
+#'
+#' set_style_options(save_path = dirname(workbook_file),
+#'                   file      = basename(workbook_file))
+#'
+#' create_table_of_contents(wb)
+#'
+#' # Manual cleanup for example
+#' unlink(workbook_file)
+#'
 #' # Reset the global options afterwards
 #' set_print(TRUE)
 #' set_output("excel")
 #'
 #' @export
 create_table_of_contents <- function(wb,
-                                     style               = excel_output_style(),
+                                     style               = .qol_options[["excel_style"]],
                                      toc_header          = "Table of Contents",
                                      toc_sheet_name      = "Contents",
                                      titles              = c(),
@@ -2078,7 +2089,6 @@ create_table_of_contents <- function(wb,
                                      subheader_colors    = c(),
                                      subheader_underline = FALSE,
                                      colored_tabs        = FALSE,
-                                     file                = NULL,
                                      print               = .qol_options[["print"]]){
     print_start_message()
 
@@ -2381,24 +2391,50 @@ create_table_of_contents <- function(wb,
     if (print){
         print_step("MAJOR", "Exporting Excel Workbook")
 
-        if (!is.null(file)){
-            file <- macro(file)
+        # Apply macros to save path and file
+        if (!is.null(style[["save_path"]])){
+            style[["save_path"]] <- macro(style[["save_path"]])
         }
 
-        if (is.null(file)){
-            if (interactive()){
-                wb$open()
+        if (!is.null(style[["file"]])){
+            style[["file"]] <- macro(style[["file"]])
+        }
+
+        # Check if only save path or file name is specified. If only one is specified
+        # print a note. Otherwise the file would not be saved and opened without a
+        # hint to why the file wasn't saved.
+        if (is.null(style[["save_path"]]) + is.null(style[["file"]]) == 1){
+            if (is.null(style[["save_path"]])){
+                print_message("NOTE", c("No save path specified. Both save path and file name with extension",
+                                        "need to be specified in the global options or style parameter for",
+                                        "the file to be saved. File won't be saved."))
+            }
+            else{
+                print_message("NOTE", c("No file name specified. Both save path and file name with extension",
+                                        "need to be specified in the global options or style parameter for",
+                                        "the file to be saved. File won't be saved."))
             }
         }
-        else if (!dir.exists(dirname(file)) || dirname(file) == "."){
-            print_message("WARNING", "Directory '[file_name]' does not exist. File won't be saved.", file_name = dirname(file))
 
+        # If no save path or file provided just open workbook
+        if (is.null(style[["save_path"]]) || is.null(style[["file"]])){
             if (interactive()){
                 wb$open()
             }
         }
         else{
-            wb$save(file = file, overwrite = TRUE)
+            # If save path doesn't exist, just open workbook
+            if (!file.exists(style[["save_path"]])){
+                print_message("WARNING", "Path does not exist: [style]", style = style[["save_path"]])
+
+                if (interactive()){
+                    wb$open()
+                }
+            }
+            # Save file
+            else{
+                wb$save(file = paste0(style[["save_path"]], "/", style[["file"]]), overwrite = TRUE)
+            }
         }
     }